# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProVentas",
        "description": "Documentación API Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear producto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "422": {
                        "description": "Validación"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener producto",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar producto",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar producto",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
        "/api/product-images/{id}": {
            "get": {
                "tags": [
                    "ProductImages"
                ],
                "summary": "Obtener imagen",
                "operationId": "0d3901c01195fd9613f9d335cd4529c7",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductImages"
                ],
                "summary": "Actualizar imagen",
                "operationId": "3486d478bac22994c0429bfcd987f5e7",
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductImages"
                ],
                "summary": "Eliminar imagen",
                "operationId": "0ed26c564cc84e973785ac02239afbd9",
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}